BG96 interfacing with Atmega2560 mictronntroller
In this project I have interfaced BG96 LTE module with Atmega2560 it will send the data to mobile number or to a server. There is another option to take data on the server i.e. using the ESP8266-12E module. This module also interfaced with the same microcontroller. The data will be read from the RFID module. All the programs are written in Embedded C language